requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
pygame>=2.5.0
numpy>=1.24.0
